Move streaming_validate to validate
