+++ lib/csvlint.rb
@@ -14,7 +14,6 @@ require 'uri_template'
 require 'csvlint/error_message'
 require 'csvlint/error_collector'
 require 'csvlint/validate'
-require 'csvlint/streaming_validate'
 require 'csvlint/wrapped_io'
 require 'csvlint/field'
